+++ packages/core/src/index.ts
@@ -45,4 +45,5 @@ export {
   ToolPluginRegistry,
 };
 
+export * from './constants';
 export * from './types';